Application_PlantCare/backend/requirements.txt
paho-mqtt==1.6.1
pymongo==4.10.1
apscheduler==3.10.4
supervisor==4.2.5
websockets==13.1